Rename sln/proj/folder
