+++ windowsforms/graphics/WinFormGraphics/MainForm.resx
