+++ windowsforms/graphics/WinFormGraphics/Properties/AssemblyInfo.cs
