+++ windowsforms/graphics/WinFormGraphics/Properties/Resources.resx
